015: implement scan method

+++ lib/OfoStepKeyboard/src/KeyboardInterface.h
@@ -9,8 +9,8 @@ class KeyboardInterface {
 public:
     ~KeyboardInterface()=default;
     virtual bool getKeys() = 0;
-    virtual void setDebounceTime(uint8_t) = 0;
-    virtual void setHoldTime(uint8_t) = 0;
+    virtual void setDebounceTime(unsigned int) = 0;
+    virtual void setHoldTime(unsigned int) = 0;
 
 };
 #endif //SEQUENCEALL_KEYBOARDINTERFACE_H
